requirements.txt
langchain>=0.2.11
langchain-openai>=0.1.17
langchain-core>=0.2.38
python-dotenv>=1.0.1
pyyaml>=6.0.2
requests>=2.32.3
tenacity>=8.5.0
